# BackEnd/openapi.yaml
openapi: 3.1.0
info:
  title: Portal institucional SENAPPEN API
  version: 1.0.0
  description: |
    API HTTP do portal. Endpoints privados usam o cookie HttpOnly de sessão configurado
    pelo servidor. Em operações que alteram dados, envie também o valor do cookie CSRF
    no cabeçalho X-CSRF-Token. A origem permitida é definida por FRONTEND_ORIGIN.
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
paths:
...
  /api/contracts:
    get:
      tags: [Contratos]
      summary: Lista contratos armazenados no SQLite
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: type, in: query, schema: { type: string } }
        - { name: year, in: query, schema: { type: string, pattern: '^\d{4}$' } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Contratos ordenados pela data mais recente
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Contratos]
      summary: Cria um contrato com anexo
      description: Multipart/form-data; anexo PDF, DOC, DOCX, JPG ou PNG até 10 MiB.
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/ContractInput'
                - type: object
                  required: [file]
                  properties:
                    file: { type: string, format: binary }
      responses:
        '201':
          description: Contrato criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Número de contrato já cadastrado }
        '413': { description: Anexo excede o limite permitido }
  /api/contracts/count:
    get:
      tags: [Contratos]
      summary: Conta contratos, com filtro anual opcional
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: year, in: query, schema: { type: string, pattern: '^\d{4}$' } }
      responses:
        '200':
          description: Total de contratos
          content:
            application/json:
              schema:
                type: object
                properties: { count: { type: integer } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/contracts/groupby/type:
    get:
      tags: [Contratos]
      summary: Conta contratos agrupados por tipo
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Contagens por tipo }
  /api/contracts/check:
    get:
      tags: [Contratos]
      summary: Verifica se um número de contrato já existe
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: number, in: query, schema: { type: string, maxLength: 100 } }
      responses:
        '200':
          description: Resultado da consulta
          content:
            application/json:
              schema:
                type: object
                properties: { exists: { type: boolean } }
  /api/contracts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Contratos]
      summary: Obtém um contrato
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Dados do contrato }
        '404': { description: Contrato não encontrado }
    put:
      tags: [Contratos]
      summary: Atualiza um contrato e, opcionalmente, seu anexo
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/ContractInput'
                - type: object
                  properties:
                    file: { type: string, format: binary }
      responses:
        '200': { description: Contrato atualizado }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { description: Contrato não encontrado }
        '409': { description: Número de contrato já cadastrado }
    delete:
      tags: [Contratos]
      summary: Exclui contrato e anexo (somente administrador)
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '200': { description: Contrato excluído }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Contrato não encontrado }
  /api/contracts/{id}/download:
    get:
      tags: [Contratos]
      summary: Baixa o anexo do contrato
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Conteúdo do anexo, como stream binário }
        '404': { description: Contrato ou anexo não encontrado }
  /api/contracts/{id}/view:
    get:
      tags: [Contratos]
      summary: Visualiza o anexo do contrato, quando suportado
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Conteúdo do anexo, como stream binário }
        '404': { description: Contrato ou anexo não encontrado }
  /api/contracts/sync:
    post:
      tags: [Contratos]
      summary: Remove metadados cujos anexos não existem (somente administrador)
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '200': { description: Sincronização concluída }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/contracts/clean-all:
    delete:
      tags: [Contratos]
      summary: Remove contratos e anexos (somente administrador)
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '200': { description: Contratos removidos }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: senappen_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    ContractInput:
      type: object
      required: [type, number, date]
      properties:
        type:
          type: string
          enum: [aditivo, servicos, fornecimento, convenio, outro]
        number: { type: string, maxLength: 100, pattern: '^[a-zA-Z0-9_-]+$' }
        date: { type: string, format: date }
        description: { type: string, maxLength: 5000 }
    Contract:
      allOf:
        - $ref: '#/components/schemas/ContractInput'
        - type: object
          required: [id, fileName, originalName, filePath, mimeType, createdAt]
          properties:
            id: { type: string, format: uuid }
            fileName: { type: string }
            originalName: { type: string }
            filePath: { type: string }
            mimeType: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time, nullable: true }
    Error:
      type: object
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
        message: { type: string }
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Sessão ausente, inválida ou expirada
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Permissão insuficiente ou token CSRF inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }